@@ -1,23 +1,17 @@
 ﻿using Blackbird.Applications.Sdk.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Apps.Taus
+namespace Apps.Taus;
+
+public class TausApplication : IApplication
 {
-    public class TausApplication : IApplication
+    public string Name
     {
-        public string Name
-        {
-            get => "TAUS";
-            set { }
-        }
+        get => "TAUS";
+        set { }
+    }
 
-        public T GetInstance<T>()
-        {
-            throw new NotImplementedException();
-        }
+    public T GetInstance<T>()
+    {
+        throw new NotImplementedException();
     }
 }
