@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Apps.Taus.Models;
 
-namespace Apps.Taus.Models
+public class Segment
 {
-    public class Segment
-    {
-        public string Value { get; set; }
-        public string Language { get; set; }
-    }
+    public string Value { get; set; }
+
+    public string Language { get; set; }
+    
+    public string? Label { get; set; }
 }
